Finished the voting system
Added the ability for players to call a vote if they have some left. And
added a votePassing prefab that each of the shrines need to have so that
they can given votes to the players. Last I added a public vote passer
paramater for the shrines (not evil shrines) so that we can give votes
to the players